refactoring - Layout Header component
